+++ test/Unit/VersionConstraintNormalizerTest.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace Localheinz\Composer\Json\Normalizer\Test\Unit;
 
+use Ergebnis\Json\Normalizer\Json;
 use Localheinz\Composer\Json\Normalizer\VersionConstraintNormalizer;
-use Localheinz\Json\Normalizer\Json;
 
 /**
  * @internal