+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`1.0.2...master`][1.0.2...master].
 
+### Changed
+
+* Started using `ergebnis/json-normalizer` instead of `localheinz/json-normalizer` ([#44]), by [@localheinz]
+
 ### Fixed
 
 * Dropped support for PHP 7.1 ([#30]), by [@localheinz]
@@ -51,5 +55,6 @@ For a full diff see [`149a393...1.0.0`][149a393...1.0.0].
 [#21]: https://github.com/localheinz/composer-json-normalizer/pull/21
 [#30]: https://github.com/localheinz/composer-json-normalizer/pull/30
 [#41]: https://github.com/localheinz/composer-json-normalizer/pull/41
+[#44]: https://github.com/localheinz/composer-json-normalizer/pull/44
 
 [@localheinz]: https://github.com/localheinz